refactor: remove all field props
