--- packages/components/forms/src/checkbox-field/README.mdx
@@ -1,62 +1,0 @@
----
-title: 'CheckboxField'
-type: 'component'
-status: 'stable'
-slug: /components/forms/src/checkbox-field/
-github: 'https://github.com/contentful/forma-36/tree/master/packages/components/forms/src/checkbox-field'
-storybook: 'https://f36-storybook.contentful.com/?path=/story/components-checkbox-field--overview'
-typescript: ./CheckboxField.tsx
----
-
-Use CheckboxField to give users the possibility to select multiple elements, zero, one or more items. This component can be used for example in forms or terms and conditions. It contains a checkbox with its label. For a single choice, please consider using [RadioButtonField component](https://f36.contentful.com/components/forms/src/radio-button-field/).
-
-## Table of contents
-
-- [How to use CheckboxField](#how-to-use-checkboxfield)
-- [Component variations](#component-variations)
-- [Content recommendations](#content-recommendations)
-- [Accessibility](#accessibility)
-- [Props](#props)
-
-## How to use CheckboxField
-
-- Checkboxes should work independently, checking one should not affect any other checkbox in the interface
-- Checkbox should always have a label
-- Listing options for a user should have some order. Some sort of system is most likely expected by the user when working with lists (numerical, time based)
-
-## Component variations
-
-- **Basic usage**
-
-```jsx
-<CheckboxField id="randomId" label="Label text" />
-```
-
-- **With helper text** - Helper text will be displayed below the label.
-
-```jsx
-<CheckboxField
-  name="checkbox name"
-  id="randomId"
-  label="Label text"
-  helpText="Helper text displayed below"
-/>
-```
-
-## Content recommendations
-
-- Start every option from a capital letter
-- Do not use commas at the end of each line
-- If the checkbox is used to agree to terms and conditions, we recommend to use first person in the sentence
-
-## Accessibility
-
-- Checkboxes should have a clear and concise label.
-- Users should be warned if selecting a checkbox will cause a change in context.
-- When the checkbox has focus, pressing the Space key changes the state of the checkbox.
-
-## Props
-
-import { Props } from '@contentful/f36-docs-utils';
-
-<Props of="CheckboxField" />
