--- packages/components/forms/src/checkbox-field/index.ts
@@ -1,2 +1,0 @@
-export { CheckboxField } from './CheckboxField';
-export type { CheckboxFieldProps } from './CheckboxField';
