+++ packages/components/forms/src/checkbox/Checkbox.test.tsx
@@ -6,21 +6,17 @@ import { Checkbox } from './Checkbox';
 
 describe('Checkbox', function () {
   it('renders the component', () => {
-    const { getByRole } = render(<Checkbox label="checkbox" />);
-
-    expect(getByRole('checkbox')).toBeTruthy();
-  });
-
-  it('renders the component with an additional class name', () => {
-    const { container } = render(
-      <Checkbox label="checkbox" className="my-extra-class" />,
+    const { getByLabelText } = render(
+      <Checkbox id="checkbox" label="label text" />,
     );
 
-    expect(container.firstChild).toHaveClass('my-extra-class');
+    const checkbox = getByLabelText('label text');
+    expect(checkbox).toBeInTheDocument();
+    expect(checkbox.id).toEqual('checkbox');
   });
 
   it('has no a11y issues', async () => {
-    const { container } = render(<Checkbox label="checkbox" />);
+    const { container } = render(<Checkbox id="checkbox" label="label text" />);
     const results = await axe(container);
 
     expect(results).toHaveNoViolations();
