--- packages/components/forms/src/checkbox/README.mdx
@@ -1,115 +1,0 @@
----
-title: 'Checkbox'
-type: 'component'
-status: 'stable'
-slug: /components/inputs/src/checkbox/
-github: 'https://github.com/contentful/forma-36/tree/master/packages/components/inputs/src/checkbox'
-storybook: 'https://f36-storybook.contentful.com/?path=/story/components-checkbox--overview'
-typescript: ./Checkbox.tsx
----
-
-Checkbox is used to show a checkbox form element. Under the hood, it uses `BaseCheckbox` with `type='checkbox'` .
-
-## Table of contents
-
-- [Table of contents](#table-of-contents)
-- [How to use Checkbox](#how-to-use-checkbox)
-- [Component variations](#component-variations)
-- [Content recommendations](#content-recommendations)
-- [Accessibility](#accessibility)
-- [Props](#props)
-
-## How to use Checkbox
-
-- You only need to define a suitable `label` to get a basic checkbox.
-- All common states (disabled, required, checked, indeterminate) are functionally and visually supported.
-- Control the current input value by setting `isChecked`.
-- Listen to native element events via `onChange`, `onBlur`, `onFocus`. More event handlers can be attached by defining them via `inputProps`.
-
-## Component variations
-
-- **default**
-
-```jsx
-// import {Checkbox} from '@contentful/f36-forms';
-
-<Checkbox label="Subscribe to newsletter" name="newsletter-subscribe" />
-```
-
-- **checked**
-
-```jsx
-// import {Checkbox} from '@contentful/f36-forms';
-
-<Checkbox
-  label="Subscribe to newsletter"
-  name="newsletter-subscribe"
-  isChecked
-/>
-```
-
-- **indeterminate**
-
-```jsx
-// import {Checkbox} from '@contentful/f36-forms';
-
-<Checkbox
-  label="Subscribe to newsletter"
-  name="newsletter-subscribe"
-  isIndeterminate
-/>
-```
-
-- **disabled**
-
-```jsx
-// import {Checkbox} from '@contentful/f36-forms';
-
-<Checkbox
-  label="Subscribe to newsletter"
-  name="newsletter-subscribe"
-  isDisabled
-/>
-```
-
-- **disabled indeterminate**
-
-```jsx
-// import {Checkbox} from '@contentful/f36-forms';
-
-<Checkbox
-  label="Subscribe to newsletter"
-  name="newsletter-subscribe"
-  isIndeterminate
-  isDisabled
-/>
-```
-
-- **disabled checked**
-
-```jsx
-// import {Checkbox} from '@contentful/f36-forms';
-
-<Checkbox
-  label="Subscribe to newsletter"
-  name="newsletter-subscribe"
-  isChecked
-  isDisabled
-/>
-```
-
-## Content recommendations
-
-- Try to use labels with short, scannable text.
-
-## Accessibility
-
-- The `label` will be rendered as `aria-label` of the input node.
-- Use a unique `id`.
-- An explicit label node is rendered for showing a custom check mark. According to [w3](https://www.w3.org/WAI/tutorials/forms/labels/#associating-labels-implicitly), explicit labels provide a better support for assistive technologies.
-
-## Props
-
-import { Props } from '@contentful/f36-docs-utils';
-
-<Props of="Checkbox" />
