+++ packages/components/forms/src/index.ts
@@ -1,6 +1,3 @@
 export * from './Label';
-export * from './checkbox-field';
-export * from './radio-button-field';
-export * from './switch-field';
+export * from './radio';
 export * from './checkbox';
-export * from './switch';
