--- packages/components/forms/src/radio-button-field/index.ts
@@ -1,2 +1,0 @@
-export { RadioButtonField } from './RadioButtonField';
-export type { RadioButtonFieldProps } from './RadioButtonField';
