--- packages/components/forms/src/radio-button/README.mdx
@@ -1,97 +1,0 @@
----
-title: 'RadioButton'
-type: 'component'
-status: 'stable'
-slug: /components/inputs/src/radio-button/
-github: 'https://github.com/contentful/forma-36/tree/master/packages/components/inputs/src/radio-button'
-storybook: 'https://f36-storybook.contentful.com/?path=/story/components-radiobutton--default'
-typescript: ./RadioButton.tsx
----
-
-RadioButton is used to show a radio button form element. Under the hood, it uses `BaseCheckbox` with `type='radio'` .
-
-## Table of contents
-
-- [Table of contents](#table-of-contents)
-- [How to use RadioButton](#how-to-use-radiobutton)
-- [Component variations](#component-variations)
-- [Content recommendations](#content-recommendations)
-- [Accessibility](#accessibility)
-- [Props](#props)
-
-## How to use RadioButton
-
-- You only need to define a suitable `label` to get a basic checkbox.
-- All common states (disabled, required, checked) are functionally and visually supported.
-- Control the current input value by setting `isChecked`.
-- Listen to native element events via `onChange`, `onBlur`, `onFocus`. More event handlers can be attached by defining them via `inputProps`.
-
-## Component variations
-
-- **default**
-
-```jsx
-// import {RadioButton} from '@contentful/f36-forms';
-
-<RadioButton
-  label="Take the blue pill"
-  name="pill-decision"
-  value="Blue pill"
-/>
-```
-
-- **checked**
-
-```jsx
-// import {RadioButton} from '@contentful/f36-forms';
-
-<RadioButton
-  label="Take the blue pill"
-  name="pill-decision"
-  value="blue"
-  isChecked
-/>
-```
-
-- **disabled**
-
-```jsx
-// import {RadioButton} from '@contentful/f36-forms';
-
-<RadioButton
-  label="Take the blue pill"
-  name="pill-decision"
-  value="blue"
-  isDisabled
-/>
-```
-
-- **disabled checked**
-
-```jsx
-// import {RadioButton} from '@contentful/f36-forms';
-
-<RadioButton
-  label="Take the blue pill"
-  name="pill-decision"
-  value="blue"
-  isChecked
-  isDisabled
-/>
-```
-
-## Content recommendations
-
-- Try to use labels with short, scannable text.
-
-## Accessibility
-
-- The `label` will be rendered as `aria-label` of the input node.
-- Use a unique `id`.
-- An explicit label node is rendered for showing a custom check mark. According to [w3](https://www.w3.org/WAI/tutorials/forms/labels/#associating-labels-implicitly), explicit labels provide a better support for assistive technologies.
-
-## Props
-
-import { Props } from '@contentful/f36-docs-utils';
-
-<Props of="RadioButton" />
