--- packages/components/forms/src/radio-button/RadioButton.md
@@ -1,7 +1,0 @@
-We're using native radio buttons (the mockup just acts as an example).
-There are 2 ways to display the labels:
-
-- Grouping a set of options with a label (medium weight). The labels for the individual radio buttons are displayed in regular weight to maintain the hierarchy.
-- Omitting the overarching label. The labels for the individual radio buttons are displayed in medium weight.
-
-![alt text](https://projects.invisionapp.com/static-signed/live-embed/118156364/217188055/1/latest/mcZDLinwhLAsHkic1FwxWQRwsPclEW1hbVp8dBr44OJ1wzXOoJPHjH9SilESFSl7NdFLltuKatfDa3gs4zXEbWrAlE/Radio-button.png 'Radio Button')
