--- packages/components/forms/src/radio-button/index.ts
@@ -1,2 +1,0 @@
-export { RadioButton } from './RadioButton';
-export type { RadioButtonProps } from './RadioButton';
