+++ packages/components/forms/src/radio/index.ts
@@ -1,0 +1,2 @@
+export { Radio } from './radio';
+export type { RadioProps } from './Radio';
