--- packages/components/forms/src/switch-field/README.mdx
@@ -1,54 +1,0 @@
----
-title: 'SwitchField'
-type: 'component'
-status: 'stable'
-slug: /components/forms/src/switch-field/
-github: 'https://github.com/contentful/forma-36/tree/master/packages/components/forms/src/switch-field'
-storybook: 'https://f36-storybook.contentful.com/?path=/story/form-elements-switchfield--basic'
-typescript: ./SwitchField.tsx
----
-
-Use SwitchField to show the users a switch toggle. It contains a checkbox input with its label. If needed the component can display an error message and hint message. For a multiple-choice, please consider using [CheckboxField component](https://f36.contentful.com/components/checkbox-field/), or the [RadioButttonField component](https://f36.contentful.com/components/radion-button-field/) for a single choice.
-
-## Table of contents
-
-- [How to use SwitchField](#how-to-use-switchfield)
-- [Component variations](#component-variations)
-- [Content recommendations](#content-recommendations)
-- [Props](#props)
-
-## How to use SwitchField
-
-- Switch should work independently, checking one should not affect any other checkbox in the interface
-- Switch should always have a label
-- Be explicit about the action that will follow if the switch is active
-
-## Component variations
-
-- **Basic usage**
-
-```jsx
-<SwitchField id="randomId" label="Label text" />
-```
-
-- **With helper text** - Example with helper text displayed
-
-```jsx
-<SwitchField
-  id="my-switch"
-  label="Some label text"
-  name="my-switch"
-  helpText="Some help text"
-/>
-```
-
-## Content recommendations
-
-- Labels should be concise and provide context.
-- Use sentence-style capitalization (only the first word in a phrase and any proper nouns capitalized)
-
-## Props
-
-import { Props } from '@contentful/f36-docs-utils';
-
-<Props of="SwitchField" />
