--- packages/components/forms/src/switch-field/index.ts
@@ -1,2 +1,0 @@
-export { SwitchField } from './SwitchField';
-export type { SwitchFieldProps } from './SwitchField';
