--- packages/components/forms/src/switch/README.mdx
@@ -1,78 +1,0 @@
----
-title: 'Switch'
-type: 'component'
-status: 'stable'
-slug: /components/inputs/src/switch/
-github: 'https://github.com/contentful/forma-36/tree/master/packages/components/inputs/src/switch'
-storybook: 'https://f36-storybook.contentful.com/?path=/story/components-switch--overview'
-typescript: ./Switch.tsx
----
-
-Switch is used to show a switch element. Under the hood, it uses `BaseCheckbox` with `type='switch'` .
-
-## Table of contents
-
-- [Table of contents](#table-of-contents)
-- [How to use Switch](#how-to-use-switch)
-- [Component variations](#component-variations)
-- [Content recommendations](#content-recommendations)
-- [Accessibility](#accessibility)
-- [Props](#props)
-
-## How to use Switch
-
-- You only need to define a suitable `label` to get a basic switch.
-- All common states (disabled, required, checked) are functionally and visually supported.
-- Control the current input value by setting `isChecked`.
-- Listen to native element events via `onChange`, `onBlur`, `onFocus`. More event handlers can be attached by defining them via `inputProps`.
-
-## Component variations
-
-- **default**
-
-```jsx
-<Switch label="Subscribe to newsletter" name="newsletter-subscribe" />
-```
-
-- **checked**
-
-```jsx
-<Switch label="Subscribe to newsletter" name="newsletter-subscribe" isChecked />
-```
-
-- **disabled**
-
-```jsx
-<Switch
-  label="Subscribe to newsletter"
-  name="newsletter-subscribe"
-  isDisabled
-/>
-```
-
-- **disabled checked**
-
-```jsx
-<Switch
-  label="Subscribe to newsletter"
-  name="newsletter-subscribe"
-  isChecked
-  isDisabled
-/>
-```
-
-## Content recommendations
-
-- Try to use labels with short, scannable text.
-
-## Accessibility
-
-- The `label` will be rendered as `aria-label` of the input node.
-- Use a unique `id`.
-- An explicit label node is rendered for showing a custom check mark. According to [w3](https://www.w3.org/WAI/tutorials/forms/labels/#associating-labels-implicitly), explicit labels provide a better support for assistive technologies.
-
-## Props
-
-import { Props } from '@contentful/f36-docs-utils';
-
-<Props of="Switch" />
