--- packages/components/forms/src/switch/index.ts
@@ -1,2 +1,0 @@
-export { Switch } from './Switch';
-export type { SwitchProps } from './Switch';
